// src/common/enums/expense-type.enum.ts
export enum Expenses {
  OFFICE = 'office',
  SHARE = 'share',
  ADVANCE = 'advance',
  INCOME = 'income',
  EXPENSE_PURCHASE = 'expense_purchase',
  EXPENSE_LOGISTICS = 'expense_logistics',
  EXPENSE_CERTIFICATES = 'expense_certificates',
  EXPENSE_OTHER = 'expense_other',
  EXPENSE_BONUS = 'expense_bonus',
  EXPENSE_MANAGER_SHARE = 'expense_manager_share'
}